requirements.txt
python-dotenv
colorama
flask
pyyaml